Added: Attack Behaviour Script; triggers an attack animation when the boss is within range of the player Hitbox script with a public attack function triggered at a certain point of the animation Added BossCanvas as a prefab which has a stand in name and stand in healthbar Updates: Implemented all changes to boss prefab Changed the names of some objects

+++ Assets/Scripts/AttackController.cs
@@ -3,7 +3,7 @@
 public class AttackController : MonoBehaviour
 {
     private Animator animator;
-    float meleeDamage = 20.0f;
+    int meleeDamage = 20;
 
     private void Start()
     {
